refactor: share is_module_loaded as utils library.
