+++ files/system/usr/libexec/secureblue/webcam_main.py
@@ -13,7 +13,10 @@ from pathlib import Path
 from typing import Final
 
 import sandbox
-from utils import ask_yes_no
+from utils import (
+    ask_yes_no,
+    is_module_loaded,
+)
 
 BLUE_HELP: Final[str] = """
 This python script toggles if webcam is enabled by creating or deleting a modprobe file at
@@ -39,16 +42,6 @@ ujust set-webcam-modules --help
 
 WEBCAM_MOD_DIR: Final[str] = "/etc/modprobe.d"
 WEBCAM_MOD_FILE: Final[str] = f"{WEBCAM_MOD_DIR}/99-disable-webcam.conf"
-
-
-def is_module_loaded(module_name: str) -> bool:
-    """Check whether the passed module name is currently loaded"""
-
-    try:
-        with open("/proc/modules", encoding="utf8") as fd:
-            return any(line.startswith(module_name + " ") for line in fd)
-    except OSError:
-        return False
 
 
 def print_status(disabled_by_file: bool) -> None:
